% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofGEM.r
\name{ofGEM}
\alias{ofGEM}
\title{A meta-analysis approach with filtering for identifying gene-level gene-environment interactions with genetic association data}
\usage{
ofGEM(Z, X, R, weight, threshold = 0.1, maxSim = 1e+06, tol = 10)
}
\arguments{
\item{Z}{the test statistics for gene-environment interactions (GxE). Each row is a SNP, and each column is a study. For multi-ethnic groups,
Z is a list with each element as a matrix for each ethnic group.}

\item{X}{the filtering statistics for GxE. Each row is a SNP, and each column is a study. For multi-ethnic groups,
x is a list with each element as a matrix for each ethnic group.}

\item{R}{the correlation matrix to simulate test and filtering statistics under the null distribution. 
The simulated test and filtering statistics are used for testing. For multi-ethnic groups,
R is a list with each element as a correlation matrix for each ethnic group.}

\item{weight}{the weight vector for each study, or the weight matrix for each SNP and each study. If the weight is common
across SNPs, it is a vector with a length equal to the number of studies. If the weight is different across SNPs, it is a matrix
with each row for a SNP and each column as a study.}

\item{threshold}{a fixed p-value threshold for filtering. The default is 0.1.}

\item{maxSim}{the maximum number of samples to be simulated for the test and filtering statistics under the null. The default is 1e6.}

\item{tol}{the tolerance number to stop the sequential sampling procedure. We count the number of simulated test statistics
with an absolute value larger than that of the calculated one based on the data for every 100 simulations. 
The sampling will stop if the count reaches tol. The default is 10. If tol = 0, the number
of simulations equals to maxSim.}
}
\value{
A list containing
\item{pval_random_mf}{the p-value based on random-effects meta-analysis with meta-filtering.}
\item{pval_fixed_mf}{the p-value based on fixed-effects meta-analysis with meta-filtering.}
\item{pval_ofGEM}{the p-value based on aggregating the p-values of fixed- and random-effects meta-analyses with meta-filtering 
using Fisher's method.}
\item{nsim}{the number of simulations that are performed.}
}
\description{
This function first conducts a meta-filtering test to filter out unpromising SNPs. It then runs a test of omnibus-filtering-based GxE meta-analysis (ofGEM) that combines the strengths of the fixed- and random-effects meta-analysis with meta-filtering. It can also analyze data from multiple ethnic groups. The p-values are calculated using a sequential sampling approach.
}
\examples{

data(sim_dat)

pval = ofGEM(Z = sim_dat$Z, X = sim_dat$X, R = sim_dat$R, weight = rep(1/6, 6))

}
\references{
Wang, Liu, Pierce, Huo, Nicolae, Olopade, Ahsan, & Chen (2017+). A meta-analysis approach with filtering for 
identifying gene-level gene-environment interactions with genetic association data. In preparation.
}
