% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_dat.r
\docType{data}
\name{sim_dat}
\alias{sim_dat}
\title{A Simulated Example data}
\value{
A list containing
\item{Z}{the simulated test statistics for GxE. Each row is a SNP, and each column is a study.}
\item{X}{the simulated filtering statistics for GxE. Each row is a SNP, and each column is a study.}
\item{R}{the simulated correlation matrix to simulate test and filtering statistics under the null distribution. 
The simulated test and filtering statistics are used for testing.}
}
\description{
This simulated data list is for demonstration.
}
\examples{
data(sim_dat)
}
