% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestplot.r
\name{forest_plot}
\alias{forest_plot}
\title{Forest plot}
\usage{
forest_plot(coef, se, sort = TRUE, exp = FALSE)
}
\arguments{
\item{coef}{the matrix for the GxE coefficients. Each row represents a SNP, and each column denotes a study. If the matrix
has colnames and/or rownames for SNPs and studies respectively. They will be shown in the forest plot. 
This is the coefficients from linear regression or generalized liear models.}

\item{se}{the matrix for the stand errors for GxE estimates. Each row represents a SNP, and each column denotes a study.}

\item{sort}{logical. If TRUE, the SNPs are ordered by the mean effect sizes across different studies. The default is TRUE.}

\item{exp}{logical. If TRUE, coef will be exponential transformed. This works for coefs obtained from logistic regressions.
The default is FALSE.}
}
\value{
A forest plot for each study ordered by SNPs.
}
\description{
It makes a forest plot for gene-environment interactions (GxE) for each study and each SNP in a gene
}
\examples{

coef = matrix(rnorm(6 * 6), 6, 6)
se = matrix(abs(runif(6 * 6, 0.1, 0.15)), 6, 6)

forest_plot(coef, se)
}
